refactor flags API for vercel
