+++ pages/api/vercel/flags.ts
@@ -1,10 +1,13 @@
+import type { NextApiRequest, NextApiResponse } from 'next';
 import { getProviderData } from 'flags/next';
-import { NextApiRequest, NextApiResponse } from 'next';
 import { verifyAccess, version, ProviderData } from 'flags';
-import * as appFlags from '../../flags';
+import * as appFlags from '../../app-flags';
 
-const handler = async (req: NextApiRequest, res: NextApiResponse) => {
-  const authorized = await verifyAccess(req.headers.authorization);
+export default async function handler(
+  req: NextApiRequest,
+  res: NextApiResponse
+) {
+  const authorized = await verifyAccess(req.headers.authorization ?? '');
   if (!authorized) {
     res.status(401).json(null);
     return;
@@ -13,6 +16,4 @@ const handler = async (req: NextApiRequest, res: NextApiResponse) => {
   const data = (await getProviderData(appFlags)) as ProviderData;
   res.setHeader('x-flags-sdk-version', version);
   res.status(200).json(data);
-};
-
-export default handler;
+}
